+++ src/jig/tests/testcase.py
@@ -42,6 +42,19 @@ def cd_gitrepo(func):
     return wrapper
 
 
+def result_with_hint(payload, hint):
+    """
+    Formats a one-line message with a hint.
+
+    This is useful for testing the output of commands that utilize the
+    :module:`jig.commands.hints` module.
+
+    :param unicode payload: the first line of the result
+    :param list hint: one of the attributes of :class:`jig.commands.hints`
+    """
+    return dedent(payload).strip() + u'\n' + u'\n'.join(hint)
+
+
 class JigTestCase(unittest.TestCase):
 
     """
@@ -93,7 +106,8 @@ class JigTestCase(unittest.TestCase):
         """
         Assert that a :py:exception:`SystemExit` has a specific exit code.
 
-        Note: this is Python 2.7/2.6 compatible since that exception changed slightly.
+        Note: this is Python 2.7/2.6 compatible since that exception changed
+        slightly.
         """
         if hasattr(exception, 'code'):
             self.assertEqual(exception.code, code)
@@ -287,7 +301,7 @@ class PluginTestCase(JigTestCase):
         section = 'plugin:test01:{0}'.format(plugindir)
         config.add_section(section)
         config.set(section, 'path',
-            join(self.fixturesdir, plugindir))
+                   join(self.fixturesdir, plugindir))
 
 
 class CommandTestCase(ViewTestCase):