+++ NEWS.rst
@@ -1,6 +1,13 @@
 News
 ====
 
+*Release 0.1.4 - March 24th, 2013*
+
+* Jig commands not include more useful messages
+* Periodically checks if installed plugins have updates and prompts to install
+  the latest plugins
+* Ignores the .jig directory when running (GitHub Issue #1)
+
 *Release 0.1.3 - February 16th, 2013*
 
 * Makes the indicator between warning messages and stop messages